Add deferred functions and make @resolve one
Currently @resolve, when not passed an RRTYPE, defaults to A,
irrespective of what the current domain is. This causes code such as:
    domain ip6 ... saddr @resolve(example.org) ...
to output an ip6tables rule with an IPv4 address.

Furthermore, dual-stacking is currently not possible; i.e.
   domain (ip ip6) ... saddr @resolve(example.org) ...
returns an iptables rule with the IPv4 address, and no ip6tables rules
at all.

Unfortunately, this is quite tricky to resolve (see #38 for an earlier
naive approach) due to the way @resolve is currently hooked in the
tokenizer. To properly address this, add a new ferm internal called
"deferred functions" which defer execution until later in parsing, when
the execution domain is known, and move @resolve there.