Ignore NXDOMAIN when resolving names with @resolve
When @resolve'ing names, ferm currently stops execution if an NXDOMAIN
(domain does not exist) is received. This is problematic because a
domain accidentally disappearing breaks the entire configuration. It is
also limiting when one wants to e.g. @resolve a host in a dual-domain
stanza, and a name only has an A or AAAA.

Note that since e89ef18 ferm ignores empty RRSETs with NOERROR, which
are really equivalent to NXDOMAINs nowadays and in fact issued instead
of NXDOMAINs in some cases (cf. RFC draft-valsorda-dnsop-black-lies-00).
This inconsistency is implicitly fixed by this patch.